src/AIEngine.ts
export class AIEngine {
    energyModels: string[];  // List of energy models used for forecasting and optimization

    constructor(models: string[]) {
        this.energyModels = models;  // Initialize with a list of models
    }

    // Method to generate energy forecast based on the models
    generateEnergyForecast(): void {
        console.log("Generating energy forecast using the following models:");

        // Loop through the energy models and simulate generating a forecast
        for (let model of this.energyModels) {
            console.log(`Using model: ${model}`);
            // Simulate generating a forecast
            const forecast = Math.random() * 100; // Random forecast for demonstration
            console.log(`Forecast for ${model}: ${forecast.toFixed(2)} units of energy.`);
        }
    }

    // Method to optimize energy usage based on forecasts and models
    optimizeEnergyUsage(): void {
        console.log("Optimizing energy usage using the following models:");

        // Loop through the energy models and simulate optimizing energy usage
        for (let model of this.energyModels) {
            console.log(`Optimizing using model: ${model}`);
            // Simulate optimization (for example, optimizing energy usage by a percentage)
            const optimizedUsage = Math.random() * 50;  // Random optimized usage for demonstration
            console.log(`Optimized energy usage for ${model}: ${optimizedUsage.toFixed(2)} units.`);
        }
    }
}

// Example usage:
const aiEngine = new AIEngine(["Model A", "Model B", "Model C"]);
aiEngine.generateEnergyForecast();
aiEngine.optimizeEnergyUsage();
